Tear down queues that have tasks at the end of each test
Currently the test teardown() stops the GRPC server, but this does
not appear to stop the actual task queues from executing. Therefore
tasks that failed in previous tests continue to retry through later
tests.

This results in confusing log output (e.g. "Task exec error with
status 405") during later tests. It also could in theory mean
unexpected calls to the localhost:5000 endpoints during later tests
(currently none of these fail in a way that matters to the current
sequence of test execution / assertion).

This commit ensures that all tests that create a task in a queue
explicitly delete the queue (which internally stops all tasks)
before stopping the emulator endpoint.

A better / less verbose solution would probably be a `destroy()`
method on the Server itself that performs a graceful shutdown
and deletes all queues & tasks, but my go is not good enough to
figure that out.