Fix panics when tests fail due to lack of http request
Previously, tests were making assertions against the detail of
the http request received even if we'd already found that the
request was nil (presumably because the request hadn't fired yet).

This was causing the test suite to panic on segfault, rather than
failing gracefully and proceeding to the next test. Instead, only
assert the details of the request if one has been captured.