Refactor tests to wait for the http request rather than fixed times
I was seeing a lot of flickering test failures if Actions was running
slow. Instead of waiting a fixed period to see an http request,
use a blocking channel to wait exactly until the test HTTP endpoint
has received the request (with a very short sleep to allow it to
send the response).

This should be more robust long-term and makes it easier to run
specific assertions against the requests without having to add
lots of capturing code at the top of the test.