Add a server option to make `PurgeQueue` perform a hard reset
With this option enabled at boot, the emulator's PurgeQueue call
becomes a synchronous operation that wipes the list of task names.

This mode allows e.g. test runs where multiple tests may need to
create tasks with the same name as a previous test, without having
to stop and start the emulator between tests.

It also simplifies waiting for clean state during a test setup,
because the emulator is guaranteed not to fire any
previously-queued tasks after the PurgeQueue operation returns.