Remove redundant `retry` argument in task dispatch calls
This was previously only used to trigger the no-retry behaviour
for `RunTask`, and is therefore now always true. Tasks will
always be considered for retry after a failure response.
They are only not retried if the queue retry config
prevents it e.g. if they have reached the max retries.