Clean up use of `if assert.XXX` in tests in favour of require
Now that all teardown is pre-registered with `defer` it's safe
(and cleaner) to use `require.XXX` for assertions that must pass
before we can check anything else.

Also extracts a helper to assert the details of a GRPC error
response, as later pending branches introduce more places where
this happens.