Fix HardReset to cope with currently-executing task
Instead of hoping that tasks will be cleared after a fixed period,
use a try-sleep-retry loop to wait for the task to clear.

If a long-running task has already been dispatched, `PurgeQueue`
may fail with a DEADLINE_EXCEEDED error. Calling code can retry
the method on a schedule to suit the application.