Add test to demonstrate HardReset fails with already-running task
The 'naive "sleep till it deletes" approach described above is too
naive' turns out to be correct.

Sleeping for a fixed period (of any length) still does not guarantee
tasks are soft-deleted before we remove them from the list. If a
task is currently executing when `PurgeQueue` is received, it isn't
deleted until it gets a response.

The naive implementation will consistenly panic in this scenario.